Add base page + search candidate tests
