+++ .gitignore
@@ -3,5 +3,4 @@ node_modules/
 /playwright-report/
 /playwright/.cache/
 package-lock.json
-
-
+playwright/.auth
